requirements.txt
transformers
jupyter
notebook
pandas
nltk
py-readability-metrics
evaluate
rouge-score
cython
spacy
torch
fairseq
benepar
datasets
dill==0.3.4
sciscpacy
keybert
Levenshtein
beautifulsoup